requirements.txt
setuptools==60.2.0
torch==1.10.0
torchvision==0.11.1
git+https://github.com/cornellius-gp/gpytorch.git@307ef9c7b56dee3635dd9e7d06a03bf51f66f6e1
git+https://github.com/pytorch/botorch.git@c2cb5bc6a2e61b2612ba505865935b681861298d
git+https://github.com/samuelstanton/upcycle.git@master
python-levenshtein==0.12.2
wandb==0.10.28
biopython==1.79
tabulate==0.8.9
pypdb==2.0
aiohttp==3.7.4
aioredis==1.3.0
omegaconf==2.1.0.dev26
hydra-core==1.1.0.dev6
cloudpickle==1.6.0
pickle5==0.0.11
pymoo==0.5.0
tqdm==4.62.3
pandas==1.3.0
deepchem==2.6.0.dev20211207173323
networkx==2.6.3
selfies==2.0.0